Moved caching to compile.js
Moved redis caching to compile.js
